Refactor: 스키마 파일명 변경(room-creation -> rooms)
